enable pool test
