--- x/gamm/pool-models/balancer/amm_test.go
@@ -1,57 +1,0 @@
-package balancer_test
-
-import (
-	"testing"
-
-	sdk "github.com/cosmos/cosmos-sdk/types"
-	"github.com/stretchr/testify/require"
-
-	"github.com/osmosis-labs/osmosis/v11/x/gamm/pool-models/balancer"
-	"github.com/osmosis-labs/osmosis/v11/x/gamm/pool-models/internal/test_helpers"
-)
-
-type BalancerTestSuite struct {
-	test_helpers.CfmmCommonTestSuite
-}
-
-func TestBalancerPoolParams(t *testing.T) {
-	// Tests that creating a pool with the given pair of swapfee and exit fee
-	// errors or succeeds as intended. Furthermore, it checks that
-	// NewPool panics in the error case.
-	tests := []struct {
-		SwapFee   sdk.Dec
-		ExitFee   sdk.Dec
-		shouldErr bool
-	}{
-		// Should work
-		{defaultSwapFee, defaultExitFee, noErr},
-		// Can't set the swap fee as negative
-		{sdk.NewDecWithPrec(-1, 2), defaultExitFee, wantErr},
-		// Can't set the swap fee as 1
-		{sdk.NewDec(1), defaultExitFee, wantErr},
-		// Can't set the swap fee above 1
-		{sdk.NewDecWithPrec(15, 1), defaultExitFee, wantErr},
-		// Can't set the exit fee as negative
-		{defaultSwapFee, sdk.NewDecWithPrec(-1, 2), wantErr},
-		// Can't set the exit fee as 1
-		{defaultSwapFee, sdk.NewDec(1), wantErr},
-		// Can't set the exit fee above 1
-		{defaultSwapFee, sdk.NewDecWithPrec(15, 1), wantErr},
-	}
-
-	for i, params := range tests {
-		PoolParams := balancer.PoolParams{
-			SwapFee: params.SwapFee,
-			ExitFee: params.ExitFee,
-		}
-		err := PoolParams.Validate(dummyPoolAssets)
-		if params.shouldErr {
-			require.Error(t, err, "unexpected lack of error, tc %v", i)
-			// Check that these are also caught if passed to the underlying pool creation func
-			_, err = balancer.NewBalancerPool(1, PoolParams, dummyPoolAssets, defaultFutureGovernor, defaultCurBlockTime)
-			require.Error(t, err)
-		} else {
-			require.NoError(t, err, "unexpected error, tc %v", i)
-		}
-	}
-}
