+++ x/gamm/pool-models/balancer/pool_test.go
@@ -8,9 +8,11 @@ import (
 
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	"github.com/stretchr/testify/require"
+	"github.com/stretchr/testify/suite"
 
 	"github.com/osmosis-labs/osmosis/v11/app/apptesting/osmoassert"
 	"github.com/osmosis-labs/osmosis/v11/x/gamm/pool-models/balancer"
+	"github.com/osmosis-labs/osmosis/v11/x/gamm/pool-models/internal/test_helpers"
 	"github.com/osmosis-labs/osmosis/v11/x/gamm/types"
 )
 
@@ -29,6 +31,56 @@ var (
 	wantErr         = true
 	noErr           = false
 )
+
+type BalancerTestSuite struct {
+	test_helpers.CfmmCommonTestSuite
+}
+
+func TestBalancerTestSuite(t *testing.T) {
+	suite.Run(t, new(BalancerTestSuite))
+}
+
+func TestBalancerPoolParams(t *testing.T) {
+	// Tests that creating a pool with the given pair of swapfee and exit fee
+	// errors or succeeds as intended. Furthermore, it checks that
+	// NewPool panics in the error case.
+	tests := []struct {
+		SwapFee   sdk.Dec
+		ExitFee   sdk.Dec
+		shouldErr bool
+	}{
+		// Should work
+		{defaultSwapFee, defaultExitFee, noErr},
+		// Can't set the swap fee as negative
+		{sdk.NewDecWithPrec(-1, 2), defaultExitFee, wantErr},
+		// Can't set the swap fee as 1
+		{sdk.NewDec(1), defaultExitFee, wantErr},
+		// Can't set the swap fee above 1
+		{sdk.NewDecWithPrec(15, 1), defaultExitFee, wantErr},
+		// Can't set the exit fee as negative
+		{defaultSwapFee, sdk.NewDecWithPrec(-1, 2), wantErr},
+		// Can't set the exit fee as 1
+		{defaultSwapFee, sdk.NewDec(1), wantErr},
+		// Can't set the exit fee above 1
+		{defaultSwapFee, sdk.NewDecWithPrec(15, 1), wantErr},
+	}
+
+	for i, params := range tests {
+		PoolParams := balancer.PoolParams{
+			SwapFee: params.SwapFee,
+			ExitFee: params.ExitFee,
+		}
+		err := PoolParams.Validate(dummyPoolAssets)
+		if params.shouldErr {
+			require.Error(t, err, "unexpected lack of error, tc %v", i)
+			// Check that these are also caught if passed to the underlying pool creation func
+			_, err = balancer.NewBalancerPool(1, PoolParams, dummyPoolAssets, defaultFutureGovernor, defaultCurBlockTime)
+			require.Error(t, err)
+		} else {
+			require.NoError(t, err, "unexpected error, tc %v", i)
+		}
+	}
+}
 
 // TestUpdateIntermediaryPoolAssetsLiquidity tests if `updateIntermediaryPoolAssetsLiquidity` returns poolAssetsByDenom map
 // with the updated liquidity given by the parameter
@@ -471,7 +523,7 @@ func TestGetPoolAssetsByDenom(t *testing.T) {
 
 // TestCalculateAmountOutAndIn_InverseRelationship tests that the same amount of token is guaranteed upon
 // sequential operation of CalcInAmtGivenOut and CalcOutAmtGivenIn.
-func (suite *BalancerTestSuite) TestBalancerCalculateAmountOutAndIn_InverseRelationship(t *testing.T) {
+func (suite *BalancerTestSuite) TestBalancerCalculateAmountOutAndIn_InverseRelationship() {
 	type testcase struct {
 		denomOut         string
 		initialPoolOut   int64
@@ -550,7 +602,7 @@ func (suite *BalancerTestSuite) TestBalancerCalculateAmountOutAndIn_InverseRelat
 
 	for _, tc := range testcases {
 		for _, swapFee := range swapFeeCases {
-			t.Run(getTestCaseName(tc, swapFee), func(t *testing.T) {
+			suite.T().Run(getTestCaseName(tc, swapFee), func(t *testing.T) {
 				ctx := suite.CreateTestContext()
 
 				poolAssetOut := balancer.PoolAsset{
@@ -573,7 +625,7 @@ func (suite *BalancerTestSuite) TestBalancerCalculateAmountOutAndIn_InverseRelat
 				require.NotNil(t, pool)
 
 				sut := func() {
-					suite.TestCalculateAmountOutAndIn_InverseRelationship(ctx, pool, poolAssetIn.Token.Denom, poolAssetOut.Token.Denom, tc.initialCalcOut, swapFeeDec)
+					suite.ValidateCalculateAmountOutAndIn_InverseRelationship(ctx, pool, poolAssetIn.Token.Denom, poolAssetOut.Token.Denom, tc.initialCalcOut, swapFeeDec)
 				}
 
 				balancerPool, ok := pool.(*balancer.Pool)
