+++ x/gamm/pool-models/stableswap/amm_test.go
@@ -5,15 +5,9 @@ import (
 	"testing"
 
 	sdk "github.com/cosmos/cosmos-sdk/types"
-	"github.com/stretchr/testify/require"
 
 	"github.com/osmosis-labs/osmosis/v11/app/apptesting/osmoassert"
-	"github.com/osmosis-labs/osmosis/v11/x/gamm/pool-models/internal/test_helpers"
 )
-
-type StableSwapTestSuite struct {
-	test_helpers.CfmmCommonTestSuite
-}
 
 func TestCFMMInvariantTwoAssets(t *testing.T) {
 	kErrTolerance := sdk.OneDec()
@@ -97,106 +91,5 @@ func TestCFMMInvariantMultiAssets(t *testing.T) {
 		fmt.Println(xOut2)
 		k3 := cfmmConstantMulti(test.xReserve.Sub(xOut2), test.yReserve.Add(test.yIn), test.uReserve, test.wSumSquares)
 		osmoassert.DecApproxEq(t, k2, k3, kErrTolerance)
-	}
-}
-
-func (suite *StableSwapTestSuite) Test_StableSwap_CalculateAmountOutAndIn_InverseRelationship(t *testing.T) {
-	type testcase struct {
-		denomOut         string
-		initialPoolOut   int64
-		initialWeightOut int64
-		initialCalcOut   int64
-
-		denomIn         string
-		initialPoolIn   int64
-		initialWeightIn int64
-	}
-
-	// For every test case in testcases, apply a swap fee in swapFeeCases.
-	testcases := []testcase{
-		{
-			denomOut:         "uosmo",
-			initialPoolOut:   1_000_000_000_000,
-			initialWeightOut: 100,
-			initialCalcOut:   100,
-
-			denomIn:         "ion",
-			initialPoolIn:   1_000_000_000_000,
-			initialWeightIn: 100,
-		},
-		{
-			denomOut:         "uosmo",
-			initialPoolOut:   1_000,
-			initialWeightOut: 100,
-			initialCalcOut:   100,
-
-			denomIn:         "ion",
-			initialPoolIn:   1_000_000,
-			initialWeightIn: 100,
-		},
-		{
-			denomOut:         "uosmo",
-			initialPoolOut:   1_000,
-			initialWeightOut: 100,
-			initialCalcOut:   100,
-
-			denomIn:         "ion",
-			initialPoolIn:   1_000_000,
-			initialWeightIn: 100,
-		},
-		{
-			denomOut:         "uosmo",
-			initialPoolOut:   1_000,
-			initialWeightOut: 200,
-			initialCalcOut:   100,
-
-			denomIn:         "ion",
-			initialPoolIn:   1_000_000,
-			initialWeightIn: 50,
-		},
-		{
-			denomOut:         "uosmo",
-			initialPoolOut:   1_000_000,
-			initialWeightOut: 200,
-			initialCalcOut:   100000,
-
-			denomIn:         "ion",
-			initialPoolIn:   1_000_000_000,
-			initialWeightIn: 50,
-		},
-	}
-
-	swapFeeCases := []string{"0", "0.001", "0.1", "0.5", "0.99"}
-
-	getTestCaseName := func(tc testcase, swapFeeCase string) string {
-		return fmt.Sprintf("tokenOutInitial: %d, tokenInInitial: %d, initialOut: %d, swapFee: %s",
-			tc.initialPoolOut,
-			tc.initialPoolIn,
-			tc.initialCalcOut,
-			swapFeeCase,
-		)
-	}
-
-	for _, tc := range testcases {
-		for _, swapFee := range swapFeeCases {
-			t.Run(getTestCaseName(tc, swapFee), func(t *testing.T) {
-				ctx := suite.CreateTestContext()
-
-				poolLiquidityIn := sdk.NewInt64Coin(tc.denomOut, tc.initialPoolOut)
-				poolLiquidityOut := sdk.NewInt64Coin(tc.denomIn, tc.initialPoolIn)
-				poolLiquidity := sdk.NewCoins(poolLiquidityIn, poolLiquidityOut)
-
-				swapFeeDec, err := sdk.NewDecFromStr(swapFee)
-				require.NoError(t, err)
-
-				exitFeeDec, err := sdk.NewDecFromStr("0")
-				require.NoError(t, err)
-
-				pool := createTestPool(t, poolLiquidity, swapFeeDec, exitFeeDec)
-				require.NotNil(t, pool)
-
-				suite.TestCalculateAmountOutAndIn_InverseRelationship(ctx, pool, poolLiquidityIn.Denom, poolLiquidityOut.Denom, tc.initialCalcOut, swapFeeDec)
-			})
-		}
 	}
 }
