+++ x/gamm/pool-models/stableswap/pool_test.go
@@ -1,0 +1,121 @@
+package stableswap
+
+import (
+	fmt "fmt"
+	"testing"
+
+	sdk "github.com/cosmos/cosmos-sdk/types"
+	"github.com/stretchr/testify/require"
+	"github.com/stretchr/testify/suite"
+
+	"github.com/osmosis-labs/osmosis/v11/x/gamm/pool-models/internal/test_helpers"
+)
+
+type StableSwapTestSuite struct {
+	test_helpers.CfmmCommonTestSuite
+}
+
+func TestStableSwapTestSuite(t *testing.T) {
+	suite.Run(t, new(StableSwapTestSuite))
+}
+
+func (suite *StableSwapTestSuite) Test_StableSwap_CalculateAmountOutAndIn_InverseRelationship() {
+	type testcase struct {
+		denomOut         string
+		initialPoolOut   int64
+		initialWeightOut int64
+		initialCalcOut   int64
+
+		denomIn         string
+		initialPoolIn   int64
+		initialWeightIn int64
+	}
+
+	// For every test case in testcases, apply a swap fee in swapFeeCases.
+	testcases := []testcase{
+		{
+			denomOut:         "uosmo",
+			initialPoolOut:   1_000_000_000_000,
+			initialWeightOut: 100,
+			initialCalcOut:   100,
+
+			denomIn:         "ion",
+			initialPoolIn:   1_000_000_000_000,
+			initialWeightIn: 100,
+		},
+		{
+			denomOut:         "uosmo",
+			initialPoolOut:   1_000,
+			initialWeightOut: 100,
+			initialCalcOut:   100,
+
+			denomIn:         "ion",
+			initialPoolIn:   1_000_000,
+			initialWeightIn: 100,
+		},
+		{
+			denomOut:         "uosmo",
+			initialPoolOut:   1_000,
+			initialWeightOut: 100,
+			initialCalcOut:   100,
+
+			denomIn:         "ion",
+			initialPoolIn:   1_000_000,
+			initialWeightIn: 100,
+		},
+		{
+			denomOut:         "uosmo",
+			initialPoolOut:   1_000,
+			initialWeightOut: 200,
+			initialCalcOut:   100,
+
+			denomIn:         "ion",
+			initialPoolIn:   1_000_000,
+			initialWeightIn: 50,
+		},
+		{
+			denomOut:         "uosmo",
+			initialPoolOut:   1_000_000,
+			initialWeightOut: 200,
+			initialCalcOut:   100000,
+
+			denomIn:         "ion",
+			initialPoolIn:   1_000_000_000,
+			initialWeightIn: 50,
+		},
+	}
+
+	swapFeeCases := []string{"0", "0.001", "0.1", "0.5", "0.99"}
+
+	getTestCaseName := func(tc testcase, swapFeeCase string) string {
+		return fmt.Sprintf("tokenOutInitial: %d, tokenInInitial: %d, initialOut: %d, swapFee: %s",
+			tc.initialPoolOut,
+			tc.initialPoolIn,
+			tc.initialCalcOut,
+			swapFeeCase,
+		)
+	}
+
+	for _, tc := range testcases {
+		for _, swapFee := range swapFeeCases {
+			suite.T().Run(getTestCaseName(tc, swapFee), func(t *testing.T) {
+				ctx := suite.CreateTestContext()
+
+				poolLiquidityIn := sdk.NewInt64Coin(tc.denomOut, tc.initialPoolOut)
+				poolLiquidityOut := sdk.NewInt64Coin(tc.denomIn, tc.initialPoolIn)
+				poolLiquidity := sdk.NewCoins(poolLiquidityIn, poolLiquidityOut)
+
+				swapFeeDec, err := sdk.NewDecFromStr(swapFee)
+				require.NoError(t, err)
+
+				exitFeeDec, err := sdk.NewDecFromStr("0")
+				require.NoError(t, err)
+
+				pool := createTestPool(t, poolLiquidity, swapFeeDec, exitFeeDec)
+				require.NotNil(t, pool)
+
+				suite.ValidateCalculateAmountOutAndIn_InverseRelationship(ctx, pool, poolLiquidityIn.Denom, poolLiquidityOut.Denom, tc.initialCalcOut, swapFeeDec)
+			})
+		}
+	}
+}
